StartedWithSearchByNameMethod
