+++ BankStartWeb/Services/ICustomersService.cs
@@ -4,17 +4,9 @@ namespace BankStartWeb.Services
 {
     public interface ICustomersService
     {
-        private readonly ApplicationDbContext _context;
-
-        public CustomersModel(ApplicationDbContext context)
-        {
-            _context = context;
-        }
-
-        public string SearchByName(string searchWord);
-        {
-            SearchWord = searchWord;
-        }
+        public string SearchByName();
+        
         public int SearchById();
+        
     }
 }
